Style modifications for Stage Editor page
Changed stylesheets and a bit of the html/js files to be more flexible in
width parameters and removed all absolute and relative positionings in
favor of proper margin manipulations. In order for the page to fit on
smaller window screens, the grid size was reduced to 60px blocks instead
of the original 85px.