Applied the user system to the stage editor
The stage editor controller (Django view) now uses django.contrib.auth
properly in stage editing, so instead of every stage being saved under
'johndoe', the current user that is logged in will be listed as stage
owner for that stage. If a user is not logged in, the black string ''
is used instead. Since a username that is blank is not allowed within
this app's login system, it's equivalent to an anonymous user for our
purposes.