Fixed stage editor bug when stage size is changed
Fixed a bug preventing the stage from rendering once the size is changed.
The bug was caused by a change in the variable type for the width/height
changing to unicode strings instead of integers. This broke the stage
creation funcitons, which were expecting integers, so the bug was fixed by
typecasting the width and height to integers before creating the stage.