Added ability to edit saved stages
Improved functionality of stage editor to allow previously saved stages
and save them to the same stage ID.

Loading stages into the editor was accomplished by passing in JSON to be
parsed. Django templates were used to directly send data to reduce the
number of server requests needed.

Saving stages required the Javascript code to keep track of the current
stageid (null if there isn't one). Once a stage is loaded into the editor,
the stageid is stored, and then it is passed back to the server when the
Render button is pressed. If the stageid matches one in the database, it
overwrites the stage with the new one.