Aligned elements in User Profile page
Added "inline-block" display on the "saved stages" and "recent stages"
sections on the user profile page to make them align horizontally. Also
added a scrollbar to "saved stages" if there are too many stages to fit
onto the screen, and increased the number of recently added stages that
appear to compensate for the large height of the div block.