Stage Editor styling overhaul, searchbar styling
Overhauled the jQuery Dialog windows that appear when you choose to resize
or name the stage by replacing them with bootstrap Modal windows instead.
Also made the resize button more visible (it's between the Play and Clear
buttons now).

Also, fixed the styling on the searchbar that somehow increased its height
within the stage editor, and only the stage editor. The fix itself spanned
all pages referencing it, though, by adding an id parameter to the search
bar and applying styles to that.