Added enemy to game, also game refactoring
Added an enemy sprite to the game, which triggers death upon collision.
Currently, enemies cannot be added via the stage editor (and already-saved
enemies don't load on the editor), but they will load on the game if you
fiddled around in shell to add them manually. Adding them to the stage
editor will come tomorrow, since it's late.

Also, refactored the code to streamline how assets are managed. Now,
instead of creating the classes for Player/Goal/Block/Enemy within a
for-loop while processing the assets in the system, it processes them
outside, and removes the need for checking the ID of objects within
game.js; this meant that the JSON sent by the server is modified
accordingly.

Finally, since smaller stages cause the endgame message to be clipped
within the smaller game window, replaced the Quintus message with an alert
function (might replace with a Bootstrap Modal) and a button to restart
the stage.