Autofocus on the game window on-load
Normally, you'd need to click on the game screen before controlling the
character. This commit added functionality such that the game window will
be "autofocused", meaning you can begin controlling the player immediately
after loading. It also re-focuses on the game screen when the Replay
button is clicked to restart the stage.