Allow stage editor blocks to resize with the window
Changed the coding within the init_canvas method in canvas-engine.js to
give .canvas-droppable objects width parameters based on percentage
instead of based on the current window size. However, heights do not
change with the widths to make squares, so to remedy the situation, added
an event listener that activates whenever the window is resized, to set
the height to be equal to the current width as it resizes.