Make username/logout button appear at every page
Used the context parameter 'user', which is included by default, to access
data from the user that is logged in without needing to check for it in
every view and add it to the context dictionary manually.

Note that the views that do check for this have not had them removed
because they might still be relevant in their own individual templates.