Mostly fixed the "sticky key" bug in the game
Because of a previous change to make the game over message a js alert,
users have occasionally been experiencing behavior where a key gets
"stuck" and clicking replay caused the player to move as if the button was
still pressed until it was pressed again. This was fixed by creating a
message in HTML at the bottom of the screen alongside the replay button.
This means that the "sticky key" bug does not occur unless a user
intentionally clicks outside the game window while holding down a button.